+++ test/libsolidity/smtCheckerTests/typecast/bytes_to_fixed_bytes_1.sol
@@ -11,4 +11,7 @@ contract C {
 		assert(g == 0x0001020304050607080900010203040506070809000102030405060708090001); // should hold
 	}
 }
+// ====
+// SMTEngine: all
+// SMTIgnoreCex: yes
 // ----